feat: pss-cln-rec-contract-roots-005 - Fold live-recording-target cluster into recording_lifecycle
Move live recording target planning off the public Recordings root into recording_lifecycle while keeping peer planner vocabulary on the thin contracts surface and exposing construction through recordings/wire.
