+++ pkg/wire/profiles.go
@@ -47,6 +47,7 @@ import (
 	settingswire "github.com/portpowered/infinite-you/pkg/services/operator_settings/wire"
 	"github.com/portpowered/infinite-you/pkg/services/recordings"
 	recordingscli "github.com/portpowered/infinite-you/pkg/services/recordings/transports/cli"
+	recordingswire "github.com/portpowered/infinite-you/pkg/services/recordings/wire"
 	systeminitialization "github.com/portpowered/infinite-you/pkg/services/system_initialization"
 	systeminitializationwire "github.com/portpowered/infinite-you/pkg/services/system_initialization/wire"
 	"github.com/portpowered/infinite-you/pkg/services/work"
@@ -75,7 +76,7 @@ func provideTerminalLoggerBuilder() terminalpolicy.LoggerBuilder {
 }
 
 func provideLiveRecordingTargetPlanner() recordings.LiveRecordingTargetPlanner {
-	return recordings.NewLiveRecordingTargetPlanner(
+	return recordingswire.NewLiveRecordingTargetPlanner(
 		platformclock.Real{},
 		uuid.NewString,
 		filepath.Join,
